// frontend/src/app/core/generated/circabc/model/helpArticle.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Representation of an Help Article in CIRCABC in a JSON object
 */
export interface HelpArticle { 
    id?: string;
    parentId?: string;
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    title?: { [key: string]: string; };
    /**
     * The object that is used to compile all the translations of a node into a JSON object. It is basically composed of a map with a key languaguage code and its value 
     */
    content?: { [key: string]: string; };
    lastUpdate?: string;
    author?: string;
    highlighted?: boolean;
    visitCounter?: number;
}